requirements.txt
Flask
flask_restful
gunicorn
flask-mysql
python-dotenv